Split up path conditions into individual files
