+++ src/CMakeLists.txt
@@ -1,4 +1,3 @@
-
 FIND_PACKAGE(BZip2 REQUIRED)
 FIND_PACKAGE(OpenGL REQUIRED)
 FIND_PACKAGE(Gettext REQUIRED)
